Move service interfaces to folder
